jit: derive register exit stores from value snapshots
Replace committed-register exit planning with value-state snapshots.

Exit analysis now records register materialization stores from the snapshot's register value family instead of carrying committed/speculative register lists. Pre-instruction exits use an explicit effect-visible value-state snapshot so same-instruction writes are not observed before faulting operations, while legacy register materialization ops can still advance that observable view.

Keep runtime emission on the temporary legacy bridge for now: generic target/value stores are planned, but store.value and precise regPart emission are left for the next lowering steps.

+++ src/backends/wasm/jit/ir/value-tracker.ts
@@ -13,6 +13,7 @@ import {
   type JitRegisterValueMap
 } from "#backends/wasm/jit/ir/register-prefix-values.js";
 import type { Reg32 } from "#x86/isa/types.js";
+import { i32 } from "#x86/state/cpu-state.js";
 
 export class JitValueTracker {
   private readonly locals = new Map<number, JitValue>();
@@ -65,7 +66,7 @@ export class JitValueTracker {
         );
         return true;
       case "value.const":
-        this.record(op.dst.id, { kind: "const", type: op.type, value: op.value });
+        this.record(op.dst.id, { kind: "const", type: op.type, value: i32(op.value) });
         return true;
       case "value.binary":
         this.record(op.dst.id, this.binaryValue(op));
